// src/app/interfaces/user.ts
export interface User {
  Id: string
  LoginEmail: string
  username: string
  email: string
  password: string
  new_password: string
  retyped_password: string
}

export interface Admin {
  Id: string
  LoginEmail: string
  username: string
  email: string
  password: string
  retyped_password: string
}